Memory matters. Replacing doubles by floats. Also, adding a new set of parameters that will be used for checkpointing.

+++ src/agents/rl/qlearning/QLearner.cpp
@@ -42,8 +42,8 @@ QLearner::QLearner(ALEInterface& ale, Features *features, Parameters *param) : R
 		Q.push_back(0);
 		Qnext.push_back(0);
 		//Initialize e:
-		e.push_back(vector<double>(numFeatures, 0.0));
-		w.push_back(vector<double>(numFeatures, 0.0));
+		e.push_back(vector<float>(numFeatures, 0.0));
+		w.push_back(vector<float>(numFeatures, 0.0));
 
 		nonZeroElig.push_back(vector<int>());
 	}
@@ -72,9 +72,9 @@ void QLearner::updateReplTrace(int action){
 	}
 }
 
-void QLearner::updateQValues(vector<int> &Features, vector<double> &QValues){
+void QLearner::updateQValues(vector<int> &Features, vector<float> &QValues){
 	for(int a = 0; a < numActions; a++){
-		double sumW = 0;
+		float sumW = 0;
 		for(unsigned int i = 0; i < Features.size(); i++){
 			sumW += w[a][Features[i]];
 		}
@@ -93,9 +93,9 @@ void QLearner::sanityCheck(){
 
 void QLearner::learnPolicy(ALEInterface& ale, Features *features){
 	struct timeval tvBegin, tvEnd, tvDiff;
-	vector<double> reward;
-	double elapsedTime;
-	double cumReward = 0, prevCumReward = 0;
+	vector<float> reward;
+	float elapsedTime;
+	float cumReward = 0, prevCumReward = 0;
 	unsigned int maxFeatVectorNorm = 1;
 	sawFirstReward = 0; firstReward = 1.0;
 
@@ -189,11 +189,11 @@ void QLearner::learnPolicy(ALEInterface& ale, Features *features){
 		}
 		gettimeofday(&tvEnd, NULL);
 		timeval_subtract(&tvDiff, &tvEnd, &tvBegin);
-		elapsedTime = double(tvDiff.tv_sec) + double(tvDiff.tv_usec)/1000000.0;
+		elapsedTime = float(tvDiff.tv_sec) + double(tvDiff.tv_usec)/1000000.0;
 		
-		double fps = double(ale.getEpisodeFrameNumber())/elapsedTime;
+		float fps = double(ale.getEpisodeFrameNumber())/elapsedTime;
 		printf("episode: %d,\t%.0f points,\tavg. return: %.1f,\t%d frames,\t%.0f fps\n",
-			episode + 1, cumReward - prevCumReward, (double)cumReward/(episode + 1.0),
+			episode + 1, cumReward - prevCumReward, (float)cumReward/(episode + 1.0),
 			ale.getEpisodeFrameNumber(), fps);
 		totalNumberFrames += ale.getEpisodeFrameNumber();
 		prevCumReward = cumReward;
@@ -202,9 +202,9 @@ void QLearner::learnPolicy(ALEInterface& ale, Features *features){
 }
 
 void QLearner::evaluatePolicy(ALEInterface& ale, Features *features){
-	double reward = 0;
-	double cumReward = 0; 
-	double prevCumReward = 0;
+	float reward = 0;
+	float cumReward = 0; 
+	float prevCumReward = 0;
 
 	//Repeat (for each episode):
 	for(int episode = 0; episode < numEpisodesEval; episode++){
@@ -222,7 +222,7 @@ void QLearner::evaluatePolicy(ALEInterface& ale, Features *features){
 		ale.reset_game();
 		sanityCheck();
 		
-		printf("%d, %f, %f\n", episode + 1, (double)cumReward/(episode + 1.0), cumReward-prevCumReward);
+		printf("%d, %f, %f\n", episode + 1, (float)cumReward/(episode + 1.0), cumReward-prevCumReward);
 		
 		prevCumReward = cumReward;
 	}
